+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/baremetalmachine/_list.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud baremetalmachine list",
-    is_preview=True,
 )
 class List(AAZCommand):
     """List bare metal machines in the provided resource group or subscription.