+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/baremetalmachine/_reimage.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud baremetalmachine reimage",
-    is_preview=True,
 )
 class Reimage(AAZCommand):
     """Reimage the provided bare metal machine.