+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/baremetalmachine/_restart.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud baremetalmachine restart",
-    is_preview=True,
 )
 class Restart(AAZCommand):
     """Restart the provided bare metal machine.