+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/cluster/_deploy.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud cluster deploy",
-    is_preview=True,
 )
 class Deploy(AAZCommand):
     """Deploy the cluster.