+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/kubernetescluster/_restart_node.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud kubernetescluster restart-node",
-    is_preview=True,
 )
 class RestartNode(AAZCommand):
     """Restart a targeted node of a Kubernetes cluster.