+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/rack/_list.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud rack list",
-    is_preview=True,
 )
 class List(AAZCommand):
     """List racks in the provided resource group or subscription.