+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/trunkednetwork/_list.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud trunkednetwork list",
-    is_preview=True,
 )
 class List(AAZCommand):
     """List trunked networks in the provided resource group or subscription.