+++ src/networkcloud/azext_networkcloud/aaz/latest/networkcloud/virtualmachine/_start.py
@@ -13,7 +13,6 @@ from azure.cli.core.aaz import *
 
 @register_command(
     "networkcloud virtualmachine start",
-    is_preview=True,
 )
 class Start(AAZCommand):
     """Start the provided virtual machine.